+++ lib/src/widgets/g_list_tile/g_list_tile.dart
@@ -1,0 +1,88 @@
+import 'package:flutter/material.dart';
+
+import '../g_text/g_text.dart';
+import '../g_text/g_text_variant.dart';
+
+class GListTile extends StatelessWidget {
+  const GListTile({
+    Key? key,
+    required this.title,
+    this.subtitle,
+    this.trailing,
+    this.isThreeLine = false,
+    this.dense,
+    this.contentPadding,
+    this.enabled = true,
+    this.onTap,
+    this.onLongPress,
+    this.mouseCursor,
+    this.selected = false,
+    this.focusColor,
+    this.hoverColor,
+    this.focusNode,
+    this.autofocus = false,
+    this.tileColor,
+    this.selectedTileColor,
+    this.enableFeedback,
+    this.horizontalTitleGap,
+    this.minVerticalPadding,
+    this.minLeadingWidth,
+  }) : super(key: key);
+
+  final String title;
+  final String? subtitle;
+  final Widget? trailing;
+  final bool isThreeLine;
+  final bool? dense;
+  final EdgeInsetsGeometry? contentPadding;
+  final bool enabled;
+  final void Function()? onTap;
+  final void Function()? onLongPress;
+  final MouseCursor? mouseCursor;
+  final bool selected;
+  final Color? focusColor;
+  final Color? hoverColor;
+  final FocusNode? focusNode;
+  final bool autofocus;
+  final Color? tileColor;
+  final Color? selectedTileColor;
+  final bool? enableFeedback;
+  final double? horizontalTitleGap;
+  final double? minVerticalPadding;
+  final double? minLeadingWidth;
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: GText(
+        title,
+        variant: GTextVariant.bodyText1,
+      ),
+      subtitle: subtitle != null
+          ? GText(
+              subtitle!,
+              colorBuilder: (c) => c.onBackground.withOpacity(.75),
+            )
+          : null,
+      trailing: trailing,
+      isThreeLine: isThreeLine,
+      dense: dense,
+      contentPadding: contentPadding,
+      enabled: enabled,
+      onTap: onTap,
+      onLongPress: onLongPress,
+      mouseCursor: mouseCursor,
+      selected: selected,
+      focusColor: focusColor,
+      hoverColor: hoverColor,
+      focusNode: focusNode,
+      autofocus: autofocus,
+      tileColor: tileColor,
+      selectedTileColor: selectedTileColor,
+      enableFeedback: enableFeedback,
+      horizontalTitleGap: horizontalTitleGap,
+      minVerticalPadding: minVerticalPadding,
+      minLeadingWidth: minLeadingWidth,
+    );
+  }
+}